Restructuring Match Algorithms
